+++ templates/application/src/_vendor.ts
@@ -5,9 +5,10 @@ import 'angular-sanitize';
 import 'angular-messages';
 import 'angular-aria';
 import 'angular-cookies';<% if (ngMaterial) { %>
-import 'angular-material';<% } %>
-import 'angular-ui-router';<% if (bootstrap) { %>
+import 'angular-material';<% } else if (bootstrap) { %>
+import 'angular-touch';
 import 'angular-ui-bootstrap';<% } %>
+import 'angular-ui-router';
 import 'angular-mocks';
 import 'angular-toastr';
 import 'restangular';